Improved the code!
I made the code that finds patterns in the names list (the "fake" ML) in to a method that is only called when the application is started and not every time a new name is generated as before. I also made the alphabet string a constant that is accessible from all methods in the program class since it is now used by two methods instead of one. The same thing goes for the Random object which is now used in the Main method as well, so that longer names will not be less common than short ones.